requirements.txt
torch
transformers
datasets
pandas
scikit-learn
pytesseract
opencv-python
python-telegram-bot==13.15
newspaper3k
python-dotenv
